Delete functionlooper.html - idle project
I haven’t worked on this little project for a while, and it may not be as useful due to the tick tag in the next Minecraft update, so I’ll remove it for now. If I want to work on it more in the future, then I may re-add it back to the website.